move various plotting utility code from doc/source/pyplots/* to pywt/_doc_utils.py
