+++ doc/source/pyplots/plot_2d_bases.py
@@ -1,102 +1,8 @@
 from itertools import product
 import numpy as np
 from matplotlib import pyplot as plt
-
-
-def wavedec_keys(level):
-    """Subband keys corresponding to a wavedec decomposition."""
-    approx = ''
-    coeffs = {}
-    for lev in range(level):
-        for k in ['a', 'd']:
-            coeffs[approx + k] = None
-        approx = 'a' * (lev + 1)
-        if lev < level - 1:
-            coeffs.pop(approx)
-    return list(coeffs.keys())
-
-
-def wavedec2_keys(level):
-    """Subband keys corresponding to a wavedec2 decomposition."""
-    approx = ''
-    coeffs = {}
-    for lev in range(level):
-        for k in ['a', 'h', 'v', 'd']:
-            coeffs[approx + k] = None
-        approx = 'a' * (lev + 1)
-        if lev < level - 1:
-            coeffs.pop(approx)
-    return list(coeffs.keys())
-
-
-def _box(bl, ur):
-    """(x, y) coordinates for the 4 lines making up a rectangular box.
-
-    Parameters
-    ==========
-    bl : float
-        The bottom left corner of the box
-    ur : float
-        The upper right corner of the box
-
-    Returns
-    =======
-    coords : 2-tuple
-        The first and second elements of the tuple are the x and y coordinates
-        of the box.
-    """
-    xl, xr = bl[0], ur[0]
-    yb, yt = bl[1], ur[1]
-    box_x = [xl, xr,
-             xr, xr,
-             xr, xl,
-             xl, xl]
-    box_y = [yb, yb,
-             yb, yt,
-             yt, yt,
-             yt, yb]
-    return (box_x, box_y)
-
-
-def _2d_wp_basis_coords(shape, keys):
-    # Coordinates of the lines to be drawn by draw_2d_wp_basis
-    coords = []
-    centers = {}  # retain center of boxes for use in labeling
-    for key in keys:
-        offset_x = offset_y = 0
-        for n, char in enumerate(key):
-            if char in ['h', 'd']:
-                offset_x += shape[0] // 2**(n + 1)
-            if char in ['v', 'd']:
-                offset_y += shape[1] // 2**(n + 1)
-        sx = shape[0] // 2**(n + 1)
-        sy = shape[1] // 2**(n + 1)
-        xc, yc = _box((offset_x, -offset_y),
-                      (offset_x + sx, -offset_y - sy))
-        coords.append((xc, yc))
-        centers[key] = (offset_x + sx // 2, -offset_y - sy // 2)
-    return coords, centers
-
-
-def draw_2d_wp_basis(shape, keys, fmt='k', plot_kwargs={}, ax=None,
-                     label_levels=0):
-    """Plot a 2D representation of a WaveletPacket2D basis."""
-    coords, centers = _2d_wp_basis_coords(shape, keys)
-    if ax is None:
-        fig, ax = plt.subplots(1, 1)
-    else:
-        fig = ax.get_figure()
-    for coord in coords:
-        ax.plot(coord[0], coord[1], fmt)
-    ax.set_axis_off()
-    ax.axis('square')
-    if label_levels > 0:
-        for key, c in centers.items():
-            if len(key) <= label_levels:
-                ax.text(c[0], c[1], key,
-                        horizontalalignment='center',
-                        verticalalignment='center')
-    return fig, ax
+from pywt._doc_utils import (wavedec_keys, wavedec2_keys, draw_2d_wp_basis,
+                             draw_2d_fswavedecn_basis)
 
 
 def _2d_fswavedecn_coords(shape, levels):
